requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==4.2.23
django-admin==2.0.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
excel-base==1.0.4
idna==3.10
isoweek==1.3.3
pillow==11.2.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.4
screen==1.0.1
six==1.17.0
sqlparse==0.5.3
TimeConvert==3.0.13
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
xlwt==1.3.0
